--- projects/nextjs/src/components/PinList/index.ts
@@ -1,1 +1,0 @@
-export { default } from './index.tsx';
