--- projects/nextjs/src/components/WorldMap/index.ts
@@ -1,1 +1,0 @@
-export { default } from './index.tsx';
